ON-14179 # Add meta to response

+++ OneBlink.SDK/models/SearchResult.cs
@@ -15,15 +15,27 @@ using Newtonsoft.Json;
 
 namespace OneBlink.SDK.Model
 {
-  public abstract class SearchResult
-  {
-    public SearchMeta meta { get; set; }
-  }
+    public abstract class SearchResult
+    {
+        public SearchMeta meta
+        {
+            get; set;
+        }
+    }
 
-  public class SearchMeta
-  {
-    public int? limit { get; set; }
-    public int? offset { get; set; }
-    public int? nextOffset { get; set; }
-  }
+    public class SearchMeta
+    {
+        public long? limit
+        {
+            get; set;
+        }
+        public long? offset
+        {
+            get; set;
+        }
+        public long? nextOffset
+        {
+            get; set;
+        }
+    }
 }
